apitypes: add ScriptClass and friends
This adds ScriptClass, which represents the type of a transaction
output's pkscript. The values of this type are NOT compatible with
dcrd's txscript.ScriptClass values! Use ScriptClassFromName to get a
text representation of a ScriptClass.

Also add check the type returned by dcrd when in GetAllTxOut as a smoke
test for changes or additions to the strings returned by dcrd.

Add IsNullDataScript helper.