Purge orphaned tip blocks on startup
This checks that the dcrdata DB is on the main chain on startup
and if it is not, then it will purge blocks back until it reaches a main
chain block according to dcrd.

Uses the existing purge-n-blocks loop. rpcutils.OrphanedTipLength
compares a database to dcrd from the tip height down until a
common mainchain ancestor is found, and returns the number of
sidechain blocks to be purged.