Central server for exchange rate data
For load-balanced backends, a central server for exchange data is
needed. Dcrrates runs an ExchangeBot and takes subscriptions from
gRPC clients, pushing updates when they are received. The Dcrrates
server requires TLS, and will auto-generate a self-signed
certificate if one is neither provided or already exists in the
application directory.

There is a script, `runprotoc.sh`, to generate the gRPC definition
files from the proto file using `protoc`. This is only needed if
the proto file is modified.

ExchangeBot itself has been updated with an option to get exchange
data from dcrrates instead of directly from the APIs. Note that
the `ExchangeBot.IsUpdated` method has been removed. It was only
used in `exchanges_test.go`, and the tests have been updated to
accomodate both modes of operation.