dcrpg: change initial table columns to TIMESTAMPTZ also
Modify the dcrpg table upgrade from 3.5.5 --> 3.6.0 so that it makes
TIMESTAMPTZ columns from the int64 unix times instead of TIMESTAMP.
This spares a pointless convertion from TIMESTAMP to TIMESTAMPTZ when
the table upgrades read the 3.8.0 --> 3.9.0 step. If someone is already
running with dcrpg 3.6.0 or later, the 3.9.0 upgrade will handle the
conversion, but will not error if the columns are already TIMESTAMPTZ.
Bump copyrights.