dcrpg: upgrade all timestamp columns to timestamptz
This bumps dcrpg schema from 3.8.0 to 3.9.0.
Also set the postgres time zone to UTC in NewChainDB[WithCancel].
Maybe reindex time column of addresses and agendas table.

* dcrpg: change initial table columns to TIMESTAMPTZ also

Modify the dcrpg table upgrade from 3.5.5 --> 3.6.0 so that it makes
TIMESTAMPTZ columns from the int64 unix times instead of TIMESTAMP.
This spares a pointless convertion from TIMESTAMP to TIMESTAMPTZ when
the table upgrades read the 3.8.0 --> 3.9.0 step. If someone is already
running with dcrpg 3.6.0 or later, the 3.9.0 upgrade will handle the
conversion, but will not error if the columns are already TIMESTAMPTZ.
Bump copyrights.

* check for the pathological state of timestamp and table versions

If the previous DB is between the 3.1/3.2 and 4.0 releases (dcrpg table
versions >3.5.5 and <3.9.0), an upgrade is likely not possible IF
PostgreSQL was running in a TimeZone other than UTC. Warn loudly in
this case.

Add the functions CheckColumnDataType, CheckCurrentTimeZone, and
CheckDefaultTimeZone to db/dcrpg/tables.go.