pubsub/types: Add IsIOTimeoutErr and use when closing connections
Closing a closed websocket connection can also result in a error with
the string "i/o timeout" since closing involves a send to the client.
When closing a websocket connection in a deferred function from the
websocket handers of explorer and pubsub, recognize such errors
and do not log the message.