rpcutils: fix BlockGate
Handle a cache miss in Block() if the waiter was too slow to retrieve
the block when notified.

Both signalHash and signalHeight require both height and hash inputs
so that they can avoid extra map lookups in blockWithHash and
hashAtHeight, respectively.

Make signalHash and signalHeight lock-free.  They are not thread-
safe -- no longer run them as goroutines. Instead have them
retrieve the wait channel, remove the wait channel from the list,
and launch at goroutine just to send the height/hash to each waiter.

WaitForHeight now checks if the block is cached, in addition to
checking fetchToHeight.