main: graceful web server shutdown
listenAndServeProto now takes the shutdown context and a waitgroup.
listenAndServeProto launches a goroutine that shutsdown the web server
when the shutdown context is cancelled, and another goroutine for the
server's listener, which blocks until the server dies.
The WaitGroup now blocks in main until the server shutdown is complete
(along with all of the other goroutines using the WaitGroup).
If the server fails to listen, it requests it's own shutdown via main's
shutdown context.
To give main a chance to halt startup if the server fails to bind to a
port, a brief delay is added before listenAndServeProto returns to main.