rework block page
This restyles /block page.

Comparisons with prototype: Instead of the coinbase reward field,
which is listed below, I've included the size field on the left,
and added the fees field on the details pane. I've removed the
"Block Details" header and stole the description for the right pane,
as "Block Hashes" was not very accurate. This left the page a little
out of sync with the recently worked mempool page, so I'm certainly
open to adding it back if we can find the right description. Where
the prototype had a single table in which the transaction type was
selectable, the separate tables are left here and the table style
from #982 has been applied. This keeps everything no-js friendly.

Adds a couple of methods for custom time display to
`explorertypes.TimeDef`. The types are not mirrored in
`dbtypes.TimeDef`.

Tweaks the way live confirmation messages are calculated so that
each `newblock.confirmations` controller target can have a custom
text.

* tweak hash box. reorder table. fix and filter rate conversion

The hash box is lightened and a little vertical padding is added.

A couple of issues in the DCR->USD conversion are addressed. Now,
it will only show the conversion for blocks within that last hour,
since older figure should be converted using historical data.

* adjust padding. truncate less on count labels. simplify block hash

* lower line height where text wraps