handle alternate DNS addresses. use separate app directory.
Alternate DNS addresses for the TLS certificate can be added with
a configuration parameter `altdnsnames`. A list of the certificate's
DNS and IP addresses are printed when a certificate is created.

The default application directory name is changed to
`dcrutil.AppDataDir("dcrrates", false)`. Leading tildes are
expanded.

Log level configuration option for to the DCRRates server.

ExchangeBotConfig.MasterBot defaults to "localhost" and "7778" for host and
server if one is provided without the other.