Central server for exchange rate data
This creates a central server for exchange rate data.

For load-balanced backends, a central server for exchange data is
needed. Dcrrates runs an ExchangeBot and takes subscriptions from
gRPC clients, pushing updates when they are received. The Dcrrates
server requires TLS, and will auto-generate a self-signed
certificate if one is neither provided or already exists in the
application directory.

There is a script, `runprotoc.sh`, to generate the gRPC definition
files from the proto file using `protoc`. This is only needed if
the proto file is modified.

ExchangeBot itself has been updated with an option to get exchange
data from dcrrates instead of directly from the APIs. Note that
the `ExchangeBot.IsUpdated` method has been removed. It was only
used in `exchanges_test.go`, and the tests have been updated to
accommodate both modes of operation.

DCRRates server will log exchange updates at Info level.

Alternate DNS addresses for the TLS certificate can be added with
a configuration parameter `altdnsnames`. A list of the certificate's
DNS and IP addresses are printed when a certificate is created.

The default application directory name is changed to
`dcrutil.AppDataDir("dcrrates", false)`. Leading tildes are
expanded.

Log level configuration option for to the DCRRates server.

ExchangeBotConfig.MasterBot defaults to "localhost" and "7778" for host and
server if one is provided without the other.