rework tx page
Restyles /tx page. Compare to single-transaction, vote, and
ticket prototypes from #451.

Compared to the prototypes, tickets get an additional two fields,
`Status` and `Spending Tx`.

For transactions in mempool, the time since received in DCRData
is displayed until the block is confirmed.

Fiat values are shown and tagged with `(today)` to indicate the
conversion uses the current rate. This is different than what was
done in #1026, where conversions are only shown for blocks < 1 hour
old.

Adds a number of methods to the `explorer.TxInfo` to help clean up
the template.

Adds a new icon for regular transactions.

Refactors `tx_controller`. Drops the `DCRThings` global object.
Using actual number of tickets in ticket pool for calculating
expiration probability rather than network parameter.