dcrpg/cache: must store empty data as non-nil
When storing data in cache, ensure the stored data (rows, etc.) are
non-nil if the BlockID is also non-nil.  This means storing an empty
non-nil slice in the case of rows and utxos, and an AddressBalance
with only the Address field set.
Also, in AddressHistory, avoid calling ReduceAddressHistory when
address rows indicates no history.