wireddb/stakedb: proper synchronous ConnectBlock funcs
The synchronous BlockConnected functions (now named ConnectBlock) just
do the work without messing with the handler goroutine.
dcrdata (main) no longer launches the goroutines that listen on blockChan.
Since the goroutines no longer run, do not create the channels. Also do not
pass the channels in the chain monitor constructors. Instead, create setters
so that other packages that want to use the monitor goroutines instead of the
synchronous functions may do so.