explorer: adjust blocks tables
This makes adjustments to the tables on /blocks, /ticketpricewindows, and
the time-based block groupings pages.

Block version is shown. On smallest screen sizes, individual counts for 4
transaction types are hidden in favour of a single tx count field. /blocks has
`explorertypes.BlockBasic`, so `Version` and `AllTxs` fields of
`explorertypes.BlockInfo` graduated to fields in `BlockBasic`, which it
already embeds. Appropriate changes were made in `makeExplorerBlockBasic`,
including a change to remove an extraneous revocation counting loop. A
`TxCount` field was added to `dbtypes.BlocksGroupedInfo` as well, for the
grouped blocks listings.

JS: `humanize.formatTxDate` updated to follow up changes in #988, and
renamed to `date`. `blocklist_controller` updated for new "version" field
in websocket block.

Style changes to tables and pagination controls, mostly for
mobile-friendliness.

Also remove extra data-type attributes.