dbtypes/dcrpg: add MergeRows
Add dbtypes.MergeRows to converts a slice of non-merged
(regular addresses table row data) into a slice of merged
address rows. This involves merging rows with the same
transaction hash into a single entry by combining the signed
values. The IsFunding field of a merged transaction indicates
if the net value is positive or not, although the Value field is
an absolute value (always positive). A map of transaction hash
to *AddressRow is also returned.  MergedRows will return a
non-nil error of a merged row is detected in the input since
only non-merged rows are expected.

Add a test for MergeRows in pgblockchain_test since it is best
tested with a live DB using the project fund address.

Also change AddressTransactionsAllMerged so it only gets
valid_mainchain rows.
Change RetrieveAllMainchainAddressMergedTxns, which
actually got all rows including invalidated and side chain rows,
so that it can be restricted to valid_mainchain=true rows.
Ensure that retrieveAddressTxns only requests valid mainchain results
when requesting merged results to match the non-merged request.