explorer: rework address page
* explorer: rework address page

Restyles address page in line with #451. There was no prototype for
a general address page, but hints are taken from the prototype for
the treasury fund page.

View selector is removed and charts are always displayed if JS is
enabled. Implements a new compact style chart controls. Tables
updated to v2 styles and made mobile friendly.

`dbtypes.AddressTx` receives two new fields, `ToStake` and
`FromStake`, which indicate the fraction of input and output DCR
respectively that is stake-related. These values are displayed as
percentages on the address' page if they are non-zero.

Adds i/o counts and fiat conversion (if available) to overview data.
A new SVG-based loading icon is used for chart and list loading.

* adjust homepage tables for tweaked hashelide

* refactor SpentUnspent and Balance functions