gov: implement live vote tracking
Adds `VoteTracker`, which is triggered with new blocks to update a
cache of voting-related data. VoteTracker produces a `VoteSummary`,
which has fields customized for use in templates. There might be
some space to integrate VoteTracker with AgendaDB.

Front end: The information is incorporated at the top of the
/agendas page here, with intention to use the same data in some way
on the homepage and more. To display a graphic without using the
`charts.js` package (which is huge), a new `VoteMeter` class
displays a simple graphic with a central value displayed using the
Canvas API. It'd be good to drop `charts.js`, which is dcrdata's
largest vendor package and we barely use it. `VoteMeter` is about
6 kB unminified, compared to 160 kB for a minified `charts.js`.