explorer: tweak navbar and footer layouts
Uses a fixed height for the navbar and footer, and sets the html
element top and bottom padding to match, which makes the body
element exactly the visible portion between the navbar and footer.
Selectively hides footer elements for mobile-friendliness.

Search bar given a button and a maximum length that is enough for
any block hash. Instead of the blue outline on focus, uses a light
border instead.