cache/dcrpg: per-address History/charts data cache
This adds to the address cache all of the history that was obtained via
DB query by "TxHistoryData".
The AddressCacheItem "metrics" placeholder is changed to history, of a
new type TxHistory that contains all types of address history used by
the charts on the addess page.
(*AddressCache).HistoryChart gets the specified chart from cache.
(*AddressCache).StoreHistoryChart stores the given chart of the
specified type in cache.

(*ChainDB).TxHistoryData is modified similar to other cache-wired
functions to first check the cache. For a cache miss, the cache is held
locked while the actual DB queries are run, and the retrieved history
data is stored in cache.

dbtypes: TimeIntervals is changed to a fixed-length array, and
AllGrouping is added to the array.  const NumIntervals = 5 is set
to guarantee the expected length, and provide a const value for
cache.TxHistory type in defining its ChartsData arrays.

Also, the cache reports logged by the address cache now make a table
of hits/misses for type of cached data.