multi: more exchange data
Adds order book and candlestick data to the `exchanges` package.
Order book data is refreshed with price data. Candlestick data is
also fetched at the same time as price data, but only once a new
candlestick should be ready. `ExchangeBot` manages a cache of
JSON-encoded byte-slices of the new data, so access the data
with `QuickSticks` and `QuickDepth` rather than directly through
the `ExchangeState`.

On the front end, a new `market` page displays a market overview.