main: create mempool collector and monitor early
The mempool monitor and collector may be created before setting up the
API types and routers.  Move the construction above Insight API
construction, and use the mempool monitor as an input to NewInsightAPI.

Remove (iapi *InsightApi).UseMempoolChecker since it is not used now.

Remove (*WiredDB).UnconfirmedTxnsForAddress and
(*WiredDB).CountUnconfirmedTransactions since they are unused, excpet
to satisfy the explorer.explorerDataSourceLite, which does not need
these functions.  Remove UnconfirmedTxnsForAddress from the
explorer.explorerDataSourceLite interface.  If WiredDB needs this
functionality in the future, it should use also mempool.MempoolMonitor.