Insight root page and api redirect
Add a page for the bare /insight path.

StatusSeeOther from /insight/api to /insight/api/status

Insight API is no longer experimental

FileServer assets under named path prefix.

middlware: OriginalRequestURI for the X-Original-Request-URI header

Add prefixPath function to the template.FuncMap. This joins a path with
a prefix. It is used in insight_root.tmpl to append to the RequestURI.

Use OriginalRequestURI with /insight and /insight/api, both of which
generate paths that must use the original request URI.

Update sample-nginx.conf