insight: set limits on Insight addrs UTXOs and txns
maxInsightAddrsUTXOs limits the number of UTXOs returned by the
addrs[/{addresses}]/utxo endpoints when the {addresses} list has more
than one address. The project fund address has 263383 UTXOs at block
height 342553, and this is a ~75MB JSON payload. Set
maxInsightAddrsUTXOs to 500000.

maxInsightAddrsTxns limits the number of transactions that may be
returned by the addrs[/{addresses}]/txs endpoints when the {addresses}
list has more than one address. This limit is applied to the "to" and
"from" URL query parameters. Note that each transaction requires a
getrawtransaction RPC call to dcrd. Set maxInsightAddrsTxns to 250.