insight: set new limits on Insight API addrs endpoints
This updates the Insight API's handling of the address list provided to the
"addrs" endpoints to allow more addresses safely.

Increase default max insight addrs to 25

maxInsightAddrsUTXOs limits the number of UTXOs returned by the
addrs[/{addresses}]/utxo endpoints when the {addresses} list has more
than one address. The project fund address has 263383 UTXOs at block
height 342553, and this is a ~75MB JSON payload. Set
maxInsightAddrsUTXOs to 500000.

maxInsightAddrsTxns limits the number of transactions that may be
returned by the addrs[/{addresses}]/txs endpoints when the {addresses}
list has more than one address. This limit is applied to the "to" and
"from" URL query parameters. Note that each transaction requires a
getrawtransaction RPC call to dcrd. Set maxInsightAddrsTxns to 250.