notification: refactor dcrd notification pipeline
The block, tx, reorg, etc updates from dcrd were being routed
through a series of channels that made the code a bit tedious.
These changes are a simplification, but should reproduce the
established behavior exactly. The 2nd level channels were dropped
in favor of registered functions. There is a still a queue loop
that handles top-level synchronization, but packages' handlers no
longer requires an additional goroutine or explicit synchronization
measures.

Handlers are registered in groups. The groups themselves are run
sequentially, but the functions within a group are run async,
allowing a high degree of sychronization control.