Charts: Dual x-axis support for all historical charts
All historical charts have a button that enables switching between the 
twox-axis types namely: Block Height and Block Time. A group button at 
the chart bottom was added to enable easy swapping of either of the two 
x-axis. Since dygraph doesn't support display of multiple x-axis the two 
distinct x-axes cannot not be displayed together.

No changes to the data stored in charts cache was made to enable an 
extra x-axis to be swapped on a chart with the same y-axis data set. The 
query below allows the index of each record retrieved to be used like a 
Block Height since the query is ordered by height. Therefore no extra 
data is required when the bin value is set to blocks.

SelectBlockStats = `SELECT height, size, time, chainwork, numtx
	FROM blocks
	WHERE is_mainchain
        AND height > $1
	ORDER BY height;`

When the bin value is set to day, a pre-populated height array entry
existed.

* Add the UI axis selector buttons stimulusjs controls

* Add height axis to all historical charts

* Fix ticket price and pow difficulty height x-axis values

* Fix the coins supply chart

* Rename zoom variables to bin

* Keep support for the deprecated zoom parameter

* Delete redundant zoom_helper implementations

* Fix the broken slider zoom selector feature

* Optimize the mapKey for legacy URLs and slider zoom features