pubsub/explorer: leave the HubRelay channels opened
There are multiple goroutines that send to HubRelay.  The WebsocketHub
is the receiver, but there are other senders.
Closing the channel is not necessary to terminate the run loops.

Also, to quit the run loops, Stop() now closes quitWSHandler instead of
sending on it.  There are NO senders, only a receiver.