api/types: API can be happy even when lagging
If the DB is at node height, the age of the best block may be fairly
large, corresponding to an extremely low likelihood block interval.

If the DB is not at node height, allow the DB to lag if the age of the
best block in the DB is very recent, suggesting that the network's best
block may still be processing.