insight: add pageNum middleware and use in /txs endpoint
The pageNum URL query param was being ignored with the /txs endpoint.
This checks for the pageNum parameter and goes to the correct page of
transactions.
This also fixes the computation of the pagesTotal JSON response field,
which previously was set to the total number of transactions rather than
than the number of pages.

note that middleware guarantees pageNum>0