Update socket.io with new api
This updates update socket.io (and engine.io) to the new version with a new API.

Presently the ping interval and timeout are hard coded, although this may change:

opts := &engineio.Options{
	PingInterval: 3 * time.Second,
	PingTimeout:  5 * time.Second,
}

* middleware: add NoOrigin

The NoOrigin middleware removes any "Origin" key from the request
header. This is intended to facilitate websocket connections with a same
origin policy, such as gorilla's default.

* api/insight: create the socket.io server allowing any Origin

The default websocket transport used with socket.io has a nil
CheckOrigin function, which results in gorilla using a same origin
policy.  This sets the CheckOrigin function to permit any origin.