views: fix vote status for abandoned proposals showing in discussion
An abandoned proposal is expected to have a VoteStatus of 0, but there are
actually abandoned proposals with a VoteStatus of 1 which indicates that they
are still in discussion. In order to categorize a proposal to either In Discussion
or Abandoned, I check if the AbandonedDate is set. If AbandonedDate is not
set, it is still in discussion.