api: streaming CSV encodes
The encoding of CSV should not have been entirely in memory, instead
encoding to the http.ResponseWriter one line at a time.

Previously this had a very specific requirement from a public facing
reverse proxy to avoid multiple simultaneous requests to the same
address but it was easy to mess up, so this is the best fix.

For the csv download path, eliminate the cr query and add an optional
"/win" suffix instead to use carriage returns at the end of each line.
This change also reduces the likelihood of server misconfiguration
causing a cache miss.