views: fix ticketpool's arg to html-head template
The /ticketpool page's handler is unique in that it does not embed
CommonPageData in a parent struct, instead passing it directly as the
only page data.  This broke the execution of "html-template" from that
page since it tried to use ".CommonPageData".
This changes it to use just "." so the template execution succeeds.