ui: Upgrade to Bootstrap 5
This diff upgrades bootstrap to v5.1.1 and applies the necessary 
changes on our UI code to keep things the same. 

Highlights

Bootstrap 5 started using `start` and `end` instead of `left` and 
`right`. Implied on the following changes:
- `text-left` to `text-start`
- `text-right` to `text-end`
- `ml-n (margin-left)` to `ms-n (margin-start)`
- `mr-n (margin-right)` to `me-n (margin-end)`
- `pl-n (padding-left)` to `ps-n (padding-start)`
- `pr-n (padding-right)` to `pe-n (padding-end)`

Changes on some `form-control` class variables, which have been 
overridden to keep consistency with the current UI design:
- `$input-color: $gray-700 !default;` to `$input-color: $body-color 
!default;`
- `$input-bg:  $white !default;` to `$input-bg: $body-bg !default;`

Changes on the `nav-link` class where a `color` property was added. This 
was overridden as well.

.font-weight-* utilities renamed as .fw-*