use db instead of searchrawtransactions
db/dcrpg:

Rewrite GetAddressTransactionsRawWithSkip to use the pg db instead of
the searchrawtransactions RPC.

Remove InsightSearchRPCAddressTransactions, which was unused.

api/types:

Stop using chainjson.VinPrevOut.

Add Vin type that is similar to chainjson.Vin.  The VinPrevOut type is
not even emulated because it was almost useless: the amout was
already in the Vin struct, and the address of the prev out is not of
high importance when requesting address transactions.

Redefine AddressTxRaw with new Vin type.

work with dcrd RPC version 7.0.0 or 8.0.0