agendas page: show all agenda meters
The quorum and approval meter targets are not usable since this is in a gohtml template
loop and it makes several identically-named targets.  The controller that initializes the
page now just uses querySelectorAll and does a forEach through all of these meters,
which now have an added class to identify them.

Also a fix for Bootstrap 5.1.3 that added a 2px top border to certain table elements.