display actual vote end block and date
This uses the "NextTVI" block and "NextTVITime" in instead of
the original "voteEnd" block and "VoteEndDate" as the actual vote end
block and date for tspends that have been approved but yet to be mined.
Also, if a tspend was short-circuited and mined, the original
vote end block and date is changed to the block and date the
tspend was mined.