cmd/dcrdata: add cache headers to improve page load on select pages (#2030)
* cmd/dcrdata: add cache headers to improve page load on select pages

This PR adds the ETag, Last-Modified and Cache-Control headers on select
explorer changes to improve user experience and save resources.

The ETag and Last-Modified are reset when *explorerUI receives a new block
or mempool data. In the future, they might be reset in other call sites where
update to backend cached data are made and API endpoints or pages depending on
cached data should then use the ETagAndLastModifiedIntercept middleware.

Signed-off-by: Philemon Ukane <ukanephilemon@gmail.com>

* format main.go

Signed-off-by: Philemon Ukane <ukanephilemon@gmail.com>

---------

Signed-off-by: Philemon Ukane <ukanephilemon@gmail.com>